requirements.txt
requests
python-dotenv
influxdb-client
urllib3
